Added get current user, moved error handling from controllers to services
